+++ tests/integration/auth/test_token.py
@@ -36,21 +36,3 @@ def test_token_incorrect_username(client, test_user):
 	token_response = client.post("/auth/token", data=token_data)
 	assert token_response.status_code == 404
 	assert token_response.json() == {"detail": "User not found"}
-
-def test_refresh_valid_token(client, tokens):
-	# Test endpoint
-	refresh_token = tokens["refresh_token"]
-	response = client.post('/auth/refresh', json=refresh_token)
-	assert response.status_code == status.HTTP_200_OK
-	response_json = response.json()
-	assert response_json.get('access_token') is not None
-
-	# Check if token grants acces to protected endpoint
-	access_token = response_json["access_token"]
-	auth_header = {"Authorization": f"Bearer {access_token}"}
-	response = client.get("/user/current", headers=auth_header)
-	assert response.status_code == status.HTTP_200_OK
-
-def test_refresh_invalid(client):
-	response = client.post('/auth/refresh', json="RandomStringNotAToken")
-	assert response.status_code == status.HTTP_401_UNAUTHORIZED
